backend/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
dj-database-url==2.2.0
Django==4.2.16
djangorestframework==3.15.2
django-stubs==5.0.4
djangorestframework-stubs==3.15.0
django-cors-headers==4.4.0
dramatiq==1.17.0
prometheus_client==0.20.0
psycopg2-binary==2.9.9
redis==5.0.8
sqlparse==0.5.1
typing_extensions==4.12.2
python-dotenv==1.0.1
ruff==0.6.4
mypy==1.11.2
mypy-extensions==1.0.0


